Move newly written Transformer code into new seperate file. Remove classes SimpleDynamicEncoder, BSpanDecoder, and ResponseDecoder, which will not be used.
