+++ implementation/transformer/tsd_net.py
@@ -22,399 +22,6 @@ def toss_(p):
     return random.randint(0, 99) <= p
 
 
-def nan(v):
-    if type(v) is float:
-        return v == float('nan')
-    return np.isnan(np.sum(v.data.cpu().numpy()))
-
-
-def get_sparse_input_aug(x_input_np):
-    """
-    sparse input of
-    :param x_input_np: [T,B]
-    :return: Numpy array: [B,T,aug_V]
-    """
-    ignore_index = [0]
-    unk = 2
-    result = np.zeros((x_input_np.shape[0], x_input_np.shape[1], cfg.vocab_size + x_input_np.shape[0]),
-                      dtype=np.float32)
-    result.fill(1e-10)
-    for t in range(x_input_np.shape[0]):
-        for b in range(x_input_np.shape[1]):
-            w = x_input_np[t][b]
-            if w not in ignore_index:
-                if w != unk:
-                    result[t][b][x_input_np[t][b]] = 1.0
-                else:
-                    result[t][b][cfg.vocab_size + t] = 1.0
-    result_np = result.transpose((1, 0, 2))
-    result = torch.from_numpy(result_np).float()
-    return result
-
-
-def init_gru(gru):
-    gru.reset_parameters()
-    for _, hh, _, _ in gru.all_weights:
-        for i in range(0, hh.size(0), gru.hidden_size):
-            torch.nn.init.orthogonal_(hh[i:i + gru.hidden_size], gain=1)
-
-
-class Attn(nn.Module):
-    def __init__(self, hidden_size):
-        super(Attn, self).__init__()
-        self.hidden_size = hidden_size
-        self.attn = nn.Linear(self.hidden_size * 2, hidden_size)
-        self.v = nn.Parameter(torch.zeros(hidden_size))
-        stdv = 1. / math.sqrt(self.v.size(0))
-        self.v.data.normal_(mean=0, std=stdv)
-
-    def forward(self, hidden, encoder_outputs, mask=False, inp_seqs=None, stop_tok=None, normalize=True):
-        encoder_outputs = encoder_outputs.transpose(0, 1)  # [B,T,H]
-        attn_energies = self.score(hidden, encoder_outputs)
-        if True or not mask:
-            normalized_energy = F.softmax(attn_energies, dim=2)  # [B,1,T]
-        else:
-            mask_idx = []
-            # inp_seqs: ndarray of [T,B]
-            # inp_seqs = inp_seqs.cpu().numpy()
-            for b in range(inp_seqs.shape[1]):
-                for t in range(inp_seqs.shape[0] + 1):
-                    if t == inp_seqs.shape[0] or inp_seqs[t, b] in stop_tok:
-                        mask_idx.append(t)
-                        break
-            mask = []
-            for mask_len in mask_idx:
-                mask.append([1.] * mask_len + [0.] * (inp_seqs.shape[0] - mask_len))
-            mask = cuda_(Variable(torch.FloatTensor(mask)))  # [B,T]
-            attn_energies = attn_energies * mask.unsqueeze(1)
-            normalized_energy = F.softmax(attn_energies, dim=2)  # [B,1,T]
-
-        context = torch.bmm(normalized_energy, encoder_outputs)  # [B,1,H]
-        return context.transpose(0, 1)  # [1,B,H]
-
-    def score(self, hidden, encoder_outputs):
-        max_len = encoder_outputs.size(1)
-        H = hidden.repeat(max_len, 1, 1).transpose(0, 1)
-        energy = self.attn(torch.cat([H, encoder_outputs], 2)).tanh()  # [B,T,2H]->[B,T,H]
-        energy = energy.transpose(2, 1)  # [B,H,T]
-        v = self.v.repeat(encoder_outputs.size(0), 1).unsqueeze(1)  # [B,1,H]
-        energy = torch.bmm(v, energy)  # [B,1,T]
-        return energy
-
-
-class SimpleTransformerEncoder(nn.Module):
-    """
-    Standard implementation of a Transformer Encoder.
-    Follows:
-        - Original paper diagram: https://arxiv.org/pdf/1706.03762.pdf, Figure 1
-        - Tutorials:
-            - https://pytorch.org/tutorials/beginner/transformer_tutorial.html
-            - https://andrewpeng.dev/transformer-pytorch/
-            - https://nlp.seas.harvard.edu/2018/04/03/attention.html
-    """
-    def __init__(self, vocabulary_size, dropout):
-        super().__init__()
-
-        # TODO: Do we want to use the dropout functionality of
-        # PositionalEncoder, TransformerEncoderLayer, or both?
-
-        D_MODEL = 512
-        # d_model: The following citations from
-        # https://arxiv.org/pdf/1706.03762.pdf explains what is d_model:
-        #
-        # The encoder is composed of a stack of N = 6 identical
-        # layers. Each layer has two sub-layers. The first is a multi-head
-        # self-attention mechanism, and the second is a simple, positionwise
-        # fully connected feed-forward network.
-        #
-        # The decoder is also composed of a stack of N = 6 identical layers. In
-        # addition to the two sub-layers in each encoder layer, the decoder
-        # inserts a third sub-layer, which performs multi-head attention over
-        # the output of the encoder stack
-        #
-        # All sub-layers in the model, as well as the embedding
-        # layers, produce outputs of dimension d_model = 512.
-
-        self.model_type='TransformerEncoder'
-
-        self.embedding = nn.Embedding(vocabulary_size, embedding_dimension)
-        self.positional_encoder = PositionalEncoder(
-            d_model=D_MODEL,
-            dropout,
-            max_len=self.cfg.max_ts
-        )
-
-        self.transformer_encoder = nn.modules.TransformerEncoder(
-            nn.modules.TransformerEncoderLayer(
-                d_model=D_MODEL,  # 512 used in the original paper.
-                nhead=8,  # Multi-head attention head count. 8 used in the original paper.
-                dropout=dropout,
-            ),
-            num_layers=6  # Number of sub-encoder-layers in the encoder. 6 used in the original paper.
-        )
-
-    def forward(self, input_seqs, input_lens, hidden=None):
-        """
-        :param input_seqs: Variable of [T,B]
-            = a batch of input sequences.
-            = 2D numpy array of dimension = (Max input sequence length x batch size.)
-        :param input_lens: *numpy array* of len for each input sequence
-        """
-        embedded = self.embedding(input_seqs)
-        embedded = self.positional_encoder(embedded)
-
-        # TODO: We need to use `input_lens`, create some masks and pass them to
-        # `transformer_encoder`:
-        #   - mask: the mask for the src sequence (optional).
-        #   - src_key_padding_mask: the mask for the src keys per batch (optional).
-        #
-        # src_key_padding_mask: For the example, this looks like [False, False,
-        # False, False, False, False, False, True, True, True] where the True
-        # positions should be masked. That is input_len * [False] +
-        # (max_input_len - input_len) * [True]?
-        outputs = self.transformer_encoder(embedded)
-
-        return outputs, embedded
-
-
-class SimpleTransformerDecoder(nn.Module):
-    """
-    We are now replacing both BSpanDecoder and ResponseDecoder by this SimpleTransformerDecoder.
-    Although BSpanDecoder and ResponseDecoder are similar, BSpanDecoder has one
-    extra feature: It uses the previous encoder endput as an extra input.
-    TODO: As future work, could we implement the same trick in our Transformer version?
-    """
-    def __init__(self, vocabulary_size, dropout):
-        super().__init__()
-
-        self.model_type='TransformerDecoder'
-
-        self.positional_encoder = PositionalEncoder(d_model, dropout, self.cfg.max_ts)
-        decoder_layers = nn.modules.TransformerDecoderLayer(d_model, nhead, dim_ff)
-        self.transformer_decoder = nn.modules.TransformerDecoder(decoder_layers, n_layers)
-
-        self.dropout_rate = dropout_rate
-        self.inp_dropout = nn.Dropout(self.dropout_rate)
-
-        self.vocab = vocab
-
-
-    def forward(self, z_enc_out, u_enc_out, u_input_np, m_t_input, degree_input, last_hidden, z_input_np):
-        sparse_z_input = Variable(self.get_sparse_selective_input(z_input_np), requires_grad=False)
-
-        m_embed = self.emb(m_t_input)
-
-        # z_enc_out are stacked (along the time-step axis) belief spans (computed in BSpanDecoder)
-        # TODO: Why should this be a BSpanDecoder output, when the variable states:
-        #   - z_ = belief span
-        #   - enc_out = encoder output
-        #   - Would a better name be z_dec_out, or is this incorrect understanding of the code?
-        z_context = self.attn_z(last_hidden, z_enc_out, mask=True, stop_tok=[self.vocab.encode('EOS_Z2')],
-                                inp_seqs=z_input_np)
-
-        # u_enc_out are "user encoder output" = user inputs encoded by the Encoder (SimpleDynamicEncoder)
-        u_context = self.attn_u(last_hidden, u_enc_out, mask=True, stop_tok=[self.vocab.encode('EOS_M')],
-                                inp_seqs=u_input_np)
-
-        # TODO what is degree_input? it is coming from the data
-        transformer_in = torch.cat([m_embed, u_context, z_context, degree_input.unsqueeze(0)], dim=2)
-        # TODO: How exactly does last_hidden work?
-        gru_out, last_hidden = self.gru(transformer_in, last_hidden)
-
-
-class SimpleDynamicEncoder(nn.Module):
-    def __init__(self, input_size, embed_size, hidden_size, n_layers, dropout):
-        super().__init__()
-        self.input_size = input_size
-        self.hidden_size = hidden_size
-        self.embed_size = embed_size
-        self.n_layers = n_layers
-        self.dropout = dropout
-        self.embedding = nn.Embedding(input_size, embed_size)
-        self.gru = nn.GRU(embed_size, hidden_size, n_layers, dropout=self.dropout, bidirectional=True)
-        init_gru(self.gru)
-
-    def forward(self, input_seqs, input_lens, hidden=None):
-        """
-        forward procedure. No need for inputs to be sorted
-        :param input_seqs: Variable of [T,B]
-        :param hidden:
-        :param input_lens: *numpy array* of len for each input sequence
-        :return:
-        """
-        batch_size = input_seqs.size(1)
-        embedded = self.embedding(input_seqs)
-        embedded = embedded.transpose(0, 1)  # [B,T,E]
-        sort_idx = np.argsort(-input_lens)
-        unsort_idx = cuda_(torch.LongTensor(np.argsort(sort_idx)))
-        input_lens = input_lens[sort_idx]
-        sort_idx = cuda_(torch.LongTensor(sort_idx))
-        embedded = embedded[sort_idx].transpose(0, 1)  # [T,B,E]
-        packed = torch.nn.utils.rnn.pack_padded_sequence(embedded, input_lens)
-        outputs, hidden = self.gru(packed, hidden)
-
-        outputs, _ = torch.nn.utils.rnn.pad_packed_sequence(outputs)
-        outputs = outputs[:, :, :self.hidden_size] + outputs[:, :, self.hidden_size:]
-        outputs = outputs.transpose(0, 1)[unsort_idx].transpose(0, 1).contiguous()
-        hidden = hidden.transpose(0, 1)[unsort_idx].transpose(0, 1).contiguous()
-        return outputs, hidden, embedded
-
-
-class BSpanDecoder(nn.Module):
-    def __init__(self, embed_size, hidden_size, vocab_size, dropout_rate, vocab):
-        super().__init__()
-        self.emb = nn.Embedding(vocab_size, embed_size)
-        if cfg.use_positional_embedding:
-            self.positional_embedding = nn.Embedding(cfg.max_ts + 1, embed_size)
-            init_pos_emb = self.position_encoding_init(cfg.max_ts + 1, embed_size)
-            self.positional_embedding.weight.data = init_pos_emb
-        self.gru = nn.GRU(hidden_size + embed_size, hidden_size, dropout=dropout_rate)
-        self.proj = nn.Linear(hidden_size * 2, vocab_size)
-
-        self.attn_u = Attn(hidden_size)
-        self.proj_copy1 = nn.Linear(hidden_size, hidden_size)
-        self.proj_copy2 = nn.Linear(hidden_size, hidden_size)
-        self.dropout_rate = dropout_rate
-
-        self.inp_dropout = nn.Dropout(self.dropout_rate)
-
-        init_gru(self.gru)
-        self.vocab = vocab
-
-    def position_encoding_init(self, n_position, d_pos_vec):
-        position_enc = np.array([[pos / np.power(10000, 2 * (j // 2) / d_pos_vec) for j in range(d_pos_vec)]
-                                 if pos != 0 else np.zeros(d_pos_vec) for pos in range(n_position)])
-
-        position_enc[1:, 0::2] = np.sin(position_enc[1:, 0::2])  # dim 2i
-        position_enc[1:, 1::2] = np.cos(position_enc[1:, 1::2])  # dim 2i+1
-        return torch.from_numpy(position_enc).type(torch.FloatTensor)
-
-    def forward(self, u_enc_out, z_tm1, last_hidden, u_input_np, pv_z_enc_out, prev_z_input_np, u_emb, pv_z_emb,
-                position):
-
-        sparse_u_input = Variable(get_sparse_input_aug(u_input_np), requires_grad=False)
-
-        if pv_z_enc_out is not None:
-            context = self.attn_u(last_hidden, torch.cat([pv_z_enc_out, u_enc_out], dim=0), mask=True,
-                                  inp_seqs=np.concatenate([prev_z_input_np, u_input_np], 0),
-                                  stop_tok=[self.vocab.encode('EOS_M')])
-        else:
-            context = self.attn_u(last_hidden, u_enc_out, mask=True, inp_seqs=u_input_np,
-                                  stop_tok=[self.vocab.encode('EOS_M')])
-        embed_z = self.emb(z_tm1)
-        # embed_z = self.inp_dropout(embed_z)
-
-        if cfg.use_positional_embedding:  # defaulty not used
-            position_label = [position] * u_enc_out.size(1)  # [B]
-            position_label = cuda_(Variable(torch.LongTensor(position_label))).view(1, -1)  # [1,B]
-            pos_emb = self.positional_embedding(position_label)
-            embed_z = embed_z + pos_emb
-
-        gru_in = torch.cat([embed_z, context], 2)
-        gru_out, last_hidden = self.gru(gru_in, last_hidden)
-        # gru_out = self.inp_dropout(gru_out)
-        gen_score = self.proj(torch.cat([gru_out, context], 2)).squeeze(0)
-        # gen_score = self.inp_dropout(gen_score)
-        u_copy_score = self.proj_copy1(u_enc_out.transpose(0, 1)).tanh()  # [B,T,H]
-        # stable version of copynet
-        u_copy_score = torch.matmul(u_copy_score, gru_out.squeeze(0).unsqueeze(2)).squeeze(2)
-        u_copy_score = u_copy_score.cpu()
-        u_copy_score_max = torch.max(u_copy_score, dim=1, keepdim=True)[0]
-        u_copy_score = torch.exp(u_copy_score - u_copy_score_max)  # [B,T]
-        u_copy_score = torch.log(torch.bmm(u_copy_score.unsqueeze(1), sparse_u_input)).squeeze(
-            1) + u_copy_score_max  # [B,V]
-        u_copy_score = cuda_(u_copy_score)
-        if pv_z_enc_out is None:
-            # u_copy_score = self.inp_dropout(u_copy_score)
-            scores = F.softmax(torch.cat([gen_score, u_copy_score], dim=1), dim=1)
-            gen_score, u_copy_score = scores[:, :cfg.vocab_size], \
-                                      scores[:, cfg.vocab_size:]
-            proba = gen_score + u_copy_score[:, :cfg.vocab_size]  # [B,V]
-            proba = torch.cat([proba, u_copy_score[:, cfg.vocab_size:]], 1)
-        else:
-            sparse_pv_z_input = Variable(get_sparse_input_aug(prev_z_input_np), requires_grad=False)
-            pv_z_copy_score = self.proj_copy2(pv_z_enc_out.transpose(0, 1)).tanh()  # [B,T,H]
-            pv_z_copy_score = torch.matmul(pv_z_copy_score, gru_out.squeeze(0).unsqueeze(2)).squeeze(2)
-            pv_z_copy_score = pv_z_copy_score.cpu()
-            pv_z_copy_score_max = torch.max(pv_z_copy_score, dim=1, keepdim=True)[0]
-            pv_z_copy_score = torch.exp(pv_z_copy_score - pv_z_copy_score_max)  # [B,T]
-            pv_z_copy_score = torch.log(torch.bmm(pv_z_copy_score.unsqueeze(1), sparse_pv_z_input)).squeeze(
-                1) + pv_z_copy_score_max  # [B,V]
-            pv_z_copy_score = cuda_(pv_z_copy_score)
-            scores = F.softmax(torch.cat([gen_score, u_copy_score, pv_z_copy_score], dim=1), dim=1)
-            gen_score, u_copy_score, pv_z_copy_score = scores[:, :cfg.vocab_size], \
-                                                       scores[:,
-                                                       cfg.vocab_size:2 * cfg.vocab_size + u_input_np.shape[0]], \
-                                                       scores[:, 2 * cfg.vocab_size + u_input_np.shape[0]:]
-            proba = gen_score + u_copy_score[:, :cfg.vocab_size] + pv_z_copy_score[:, :cfg.vocab_size]  # [B,V]
-            proba = torch.cat([proba, pv_z_copy_score[:, cfg.vocab_size:], u_copy_score[:, cfg.vocab_size:]], 1)
-        return gru_out, last_hidden, proba
-
-
-class ResponseDecoder(nn.Module):
-    def __init__(self, embed_size, hidden_size, vocab_size, degree_size, dropout_rate, gru, proj, emb, vocab):
-        super().__init__()
-        self.emb = emb
-        self.attn_z = Attn(hidden_size)
-        self.attn_u = Attn(hidden_size)
-        self.gru = gru
-        init_gru(self.gru)
-        self.proj = proj
-        self.proj_copy1 = nn.Linear(hidden_size, hidden_size)
-        self.proj_copy2 = nn.Linear(hidden_size, hidden_size)
-        self.dropout_rate = dropout_rate
-
-        self.vocab = vocab
-
-    def get_sparse_selective_input(self, x_input_np):
-        result = np.zeros((x_input_np.shape[0], x_input_np.shape[1], cfg.vocab_size + x_input_np.shape[0]),
-                          dtype=np.float32)
-        result.fill(1e-10)
-        reqs = ['address', 'phone', 'postcode', 'pricerange', 'area']
-        for t in range(x_input_np.shape[0] - 1):
-            for b in range(x_input_np.shape[1]):
-                w = x_input_np[t][b]
-                word = self.vocab.decode(w)
-                if word in reqs:
-                    slot = self.vocab.encode(word + '_SLOT')
-                    result[t + 1][b][slot] = 1.0
-                else:
-                    if w == 2 or w >= cfg.vocab_size:
-                        result[t + 1][b][cfg.vocab_size + t] = 5.0
-                    else:
-                        result[t + 1][b][w] = 1.0
-        result_np = result.transpose((1, 0, 2))
-        result = torch.from_numpy(result_np).float()
-        return result
-
-    def forward(self, z_enc_out, u_enc_out, u_input_np, m_t_input, degree_input, last_hidden, z_input_np):
-        sparse_z_input = Variable(self.get_sparse_selective_input(z_input_np), requires_grad=False)
-
-        m_embed = self.emb(m_t_input)
-        z_context = self.attn_z(last_hidden, z_enc_out, mask=True, stop_tok=[self.vocab.encode('EOS_Z2')],
-                                inp_seqs=z_input_np)
-        u_context = self.attn_u(last_hidden, u_enc_out, mask=True, stop_tok=[self.vocab.encode('EOS_M')],
-                                inp_seqs=u_input_np)
-        gru_in = torch.cat([m_embed, u_context, z_context, degree_input.unsqueeze(0)], dim=2)
-        gru_out, last_hidden = self.gru(gru_in, last_hidden)
-        gen_score = self.proj(torch.cat([z_context, u_context, gru_out], 2)).squeeze(0)
-        z_copy_score = self.proj_copy2(z_enc_out.transpose(0, 1)).tanh()
-        z_copy_score = torch.matmul(z_copy_score, gru_out.squeeze(0).unsqueeze(2)).squeeze(2)
-        z_copy_score = z_copy_score.cpu()
-        z_copy_score_max = torch.max(z_copy_score, dim=1, keepdim=True)[0]
-        z_copy_score = torch.exp(z_copy_score - z_copy_score_max)  # [B,T]
-        z_copy_score = torch.log(torch.bmm(z_copy_score.unsqueeze(1), sparse_z_input)).squeeze(
-            1) + z_copy_score_max  # [B,V]
-        z_copy_score = cuda_(z_copy_score)
-
-        scores = F.softmax(torch.cat([gen_score, z_copy_score], dim=1), dim=1)
-        gen_score, z_copy_score = scores[:, :cfg.vocab_size], \
-                                  scores[:, cfg.vocab_size:]
-        proba = gen_score + z_copy_score[:, :cfg.vocab_size]  # [B,V]
-        proba = torch.cat([proba, z_copy_score[:, cfg.vocab_size:]], 1)
-        return proba, last_hidden, gru_out
-
-
 class TSD(nn.Module):
     def __init__(self, embed_size, hidden_size, vocab_size, degree_size, layer_num, dropout_rate, z_length,
                  max_ts, beam_search=False, teacher_force=100, **kwargs):
@@ -422,6 +29,8 @@ class TSD(nn.Module):
         self.vocab = kwargs['vocab']
         self.reader = kwargs['reader']
         self.emb = nn.Embedding(vocab_size, embed_size)
+
+        # TODO: Use SimpleTransformerEncoder and SimpleTransformerDecoder instead.
         self.dec_gru = nn.GRU(degree_size + embed_size + hidden_size * 2, hidden_size, dropout=dropout_rate)
         self.proj = nn.Linear(hidden_size * 3, vocab_size)
         self.u_encoder = SimpleDynamicEncoder(vocab_size, embed_size, hidden_size, layer_num, dropout_rate)
